// src/core/src/enums/change-entity.ts
/**
 * @license
 * Copyright (c) Noldova.
 *
 * This source code is licensed under the license found in the
 * LICENSE file in the root directory of this source tree.
 */

export enum ChangeEntity {
  Teammate = "Teammate",
  ConversationMember = "ConversationMember",
  Project = "Project",
  Conversation = "Conversation",
  Message = "Message",
  Approval = "Approval",
  ProviderAccount = "ProviderAccount"
}